Fix a 32-bit overflow in wanFwMark and a silent-pass gap in the carrying-flag test
Compare wan.FwMark against uint32's max through uint64 in wanFwMark, since int(^uint32(0)) is -1 on a 32-bit int target. Use the two-value map lookup for snapshot.Routing in TestPublishLiveStateReportsTheActiveTier and fail when a provider is missing, since a plain index returns MemberRouting{}'s zero value and would silently satisfy an expected-false assertion for an absent record.

+++ mwan/go/cmd/mwan/ifmgr_module_configs_linux.go
@@ -763,7 +763,7 @@ func wanFwMark(wan sharedWAN) (uint32, error) {
 	if wan.FwMark < 0 {
 		return 0, fmt.Errorf("network.json wan %s fw-mark must be >= 0", wan.Name)
 	}
-	if wan.FwMark > int(^uint32(0)) {
+	if uint64(wan.FwMark) > uint64(^uint32(0)) {
 		return 0, fmt.Errorf("network.json wan %s fw-mark %d exceeds uint32", wan.Name, wan.FwMark)
 	}
 	return uint32(wan.FwMark), nil
